requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
ephem==4.1.4
pyswisseph==2.10.3.2
kerykeion==4.26.2

# Development dependencies
pytest==7.4.4
black==23.12.1
flake8==7.0.0
isort==5.13.2
